Enable TINYINT and SMALLINT in JDBC connectors

+++ presto-mysql/src/test/java/com/facebook/presto/plugin/mysql/TestMySqlIntegrationSmokeTest.java
@@ -118,7 +118,7 @@ public class TestMySqlIntegrationSmokeTest
 
         MaterializedResult actual = computeActual("SHOW COLUMNS FROM mysql_test_tinyint1");
         MaterializedResult expected = MaterializedResult.resultBuilder(getSession(), TINYINT)
-                .row("c_tinyint", "integer", "").build();
+                .row("c_tinyint", "tinyint", "").build();
 
         assertEquals(actual, expected);
 
@@ -128,7 +128,7 @@ public class TestMySqlIntegrationSmokeTest
         MaterializedRow row = getOnlyElement(materializedRows);
 
         assertEquals(row.getFields().size(), 1);
-        assertEquals(row.getField(0), 127);
+        assertEquals(row.getField(0), (byte) 127);
 
         assertUpdate("DROP TABLE mysql_test_tinyint1");
     }
